# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teknasyon App",
        "description": "Swagger OpenApi for Teknasyon App",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Teknasyon"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "register"
                ],
                "summary": "Register new device",
                "description": "When a mobile device starts up first, it must register at the API, and device info such as uID, appID,language and operating system (OS) must be saved to the device table at registration.",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceRegistrationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "something bad has been occurred!"
                    }
                }
            }
        },
        "/purchase": {
            "post": {
                "tags": [
                    "purchase"
                ],
                "summary": "Purchase new subscription",
                "description": "It is the purchase request made in the mobile app. Mobile client sends client-token and receipt (it can be a random, meaningless hash) parameters to this API.",
                "operationId": "purchase",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PurchaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "something bad has been occurred!"
                    }
                }
            }
        },
        "/check": {
            "get": {
                "tags": [
                    "check"
                ],
                "summary": "Check Subscription",
                "description": "The mobile client can call this endpoint whenever it is on or at any step it deems necessary. It should return current subscription status as the response to your request, only with the client-token parameter.",
                "operationId": "check",
                "parameters": [
                    {
                        "name": "client_token",
                        "in": "query",
                        "description": "Unique Client Token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceRegistrationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "something bad has been occurred!"
                    }
                }
            }
        },
        "/report": {
            "get": {
                "tags": [
                    "report"
                ],
                "summary": "Generate Report",
                "description": "Reports showing new, expired and renewed subscriptions on the basis of app, day and OS.",
                "operationId": "report",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeviceRegistrationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "something bad has been occurred!"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "DeviceRegistrationResponse": {
                "title": "DeviceRegistrationResponse",
                "description": "Device model",
                "properties": {
                    "result": {
                        "title": "result",
                        "description": "Successful registration response",
                        "type": "string",
                        "example": "Ok"
                    },
                    "message": {
                        "title": "message",
                        "description": "Registration message",
                        "type": "string",
                        "example": "Device registered"
                    },
                    "client_token": {
                        "title": "client_token",
                        "description": "Client Token",
                        "type": "string",
                        "example": "6bf768c9-d576-4168-9905-a0ess9"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DeviceRegistrationResponse"
                }
            },
            "PurchaseResponse": {
                "title": "PurchaseResponse",
                "description": "Device model",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Successful Purchase",
                        "type": "string",
                        "example": "Success"
                    },
                    "message": {
                        "title": "message",
                        "description": "Registration message",
                        "type": "string",
                        "example": "Successful purchase"
                    },
                    "expire_date": {
                        "title": "expire_date",
                        "description": "Subscription expiry date",
                        "type": "string",
                        "example": "2020-01-01"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PurchaseResponse"
                }
            },
            "PurchaseRequest": {
                "title": "Purchase Request",
                "description": "Purchase Request",
                "required": [
                    "client_token"
                ],
                "properties": {
                    "client_token": {
                        "title": "client_token ",
                        "description": "Client Token",
                        "type": "string",
                        "example": "2w2ee-665t5-676ty-76bvs"
                    },
                    "receipt": {
                        "title": "receipt",
                        "description": "Hashed Receipt number",
                        "type": "string",
                        "example": "2w3e23e2233"
                    }
                },
                "type": "object"
            },
            "RegisterRequest": {
                "title": "Store Project request",
                "description": "Store Project request body data",
                "required": [
                    "u_id"
                ],
                "properties": {
                    "os": {
                        "title": "os",
                        "description": "Operating System",
                        "type": "string",
                        "example": "android"
                    },
                    "lang": {
                        "title": "lang",
                        "description": "Language",
                        "type": "string",
                        "example": "English"
                    },
                    "u_id": {
                        "title": "u_id",
                        "description": "Unique Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1232232
                    },
                    "app_id": {
                        "title": "app_id",
                        "description": "App Id",
                        "type": "integer",
                        "format": "int64",
                        "example": 123223
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Projects",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}